requirements.txt
emoji~=0.6.0
python-telegram-bot~=13.0